Datové zdroje pro Portál dobrovolníka (#283)
* Update caniuse-lite to prevent it from spamming the logs

* Join type definitions into a single file

Makes it easier to glance at the whole submodule without switching files.

* Refactor Airtable data loading

I have moved side effects (process.env reading) from the Gatsby Airtable
plugin into its config and simplified related loading code.

* Refactor some arrow functions to named functions

Let’s only use arrow functions for shorter blocks of code, the
type signatures are harder to read for them.

* Refactor node import factories to plain function calls

The factories were hard to read, with too many levels of indirection, let’s
turn them into plain (and testable!) function calls. I think the code can be
further simplified, but let’s break it into several steps.

* Further refactor node loading

The main idea here is that the createContentDigest function from
Gatsby doesn’t have to be passed from sourceNodesArgs as a dependency.
According to the docs it just performs some hashing, so it has no
state and can be called directly. This further simplifies the
conversion from our datatypes to Gatsby nodes, since the conversion
is now just a plain function of T => NodeInput type.

* Trivial refactoring: arrow functions to named functions

* Trivial refactoring: reshuffle types in files

* Trivia: make Prettier quote props consistently

* Add basic Event type and Airtable transformer

* Load Event nodes from Airtable

* Load Event–Owner relationship

* Load Event–Project relationship

This makes it possible to load a cover photo for an Event and also
cross-link to the project on our website.

* Load Event–Tag relationship

* Add mock data for events

* Refactoring: join smaller files together

* Remove GraphQL schema customization

It looks like we don’t use the feature, so let’s not pay for it
by having to update the GraphQL schema definitions.

* Fix mock data consistency

* Use real Airtable JSON dumps as mock data sources