Rychlejší build (#299)
* Add local package.json for serverless functions

When there is no package.json in api/, Vercel uses the main package.json
when deploying _each_ of the serverless functions. And since the main
package.json is quite big, this is slowing down our builds. See here for
details:

https://github.com/vercel/vercel/discussions/4995

* Delete the RSVP API endpoint

There are some type issues in there and we don’t use the endpoint
for anything meaningful at the moment, so let’s axe it.